+++ qualtran/bloqs/mcmt/__init__.py
@@ -1,0 +1,6 @@
+from qualtran.bloqs.mcmt.and_bloq import And, MultiAnd
+from qualtran.bloqs.mcmt.multi_control_multi_target_pauli import (
+    MultiControlPauli,
+    MultiControlX,
+    MultiTargetCNOT,
+)
