+++ qualtran/bloqs/reflection_using_prepare.py
@@ -21,7 +21,7 @@ from numpy.typing import NDArray
 
 from qualtran import GateWithRegisters, QBit, Register, Signature
 from qualtran._infra.gate_with_registers import merge_qubits, total_bits
-from qualtran.bloqs.multi_control_multi_target_pauli import MultiControlPauli
+from qualtran.bloqs.mcmt.multi_control_multi_target_pauli import MultiControlPauli
 from qualtran.bloqs.select_and_prepare import PrepareOracle
