+++ qualtran/bloqs/unary_iteration_bloq.py
@@ -23,8 +23,8 @@ from numpy.typing import NDArray
 
 from qualtran import GateWithRegisters, Register, Signature
 from qualtran._infra.gate_with_registers import merge_qubits, total_bits
-from qualtran.bloqs import and_bloq
 from qualtran.bloqs.basic_gates import CNOT, XGate
+from qualtran.bloqs.mcmt import and_bloq
 
 if TYPE_CHECKING:
     import sympy
