+++ qualtran/drawing/bloq_counts_graph_test.py
@@ -13,7 +13,7 @@
 #  limitations under the License.
 import re
 
-from qualtran.bloqs.and_bloq import MultiAnd
+from qualtran.bloqs.mcmt.and_bloq import MultiAnd
 from qualtran.drawing import format_counts_graph_markdown, format_counts_sigma, GraphvizCounts
 from qualtran.resource_counting import get_bloq_call_graph
