+++ qualtran/serialization/bloq.py
@@ -37,8 +37,9 @@ from qualtran import (
     Signature,
     Soquet,
 )
-from qualtran.bloqs import and_bloq, arithmetic, basic_gates, factoring, swap_network
+from qualtran.bloqs import arithmetic, basic_gates, factoring, swap_network
 from qualtran.bloqs.arithmetic import sorting
+from qualtran.bloqs.mcmt import and_bloq
 from qualtran.bloqs.util_bloqs import Allocate, ArbitraryClifford, Free, Join, Split
 from qualtran.cirq_interop import CirqGateAsBloq
 from qualtran.protos import bloq_pb2
