+++ qualtran/simulation/tensor/_flattening_test.py
@@ -11,8 +11,8 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from qualtran.bloqs.and_bloq import MultiAnd
 from qualtran.bloqs.basic_gates import CNOT
+from qualtran.bloqs.mcmt.and_bloq import MultiAnd
 from qualtran.simulation.tensor import bloq_has_custom_tensors
